3781: Introduce DragTracker and implement tool dragging in terms of that
We use introduce DragTrackerAdapter to adapt the tool controllers to the
new DragTracker interface so that we can use it immediately. We will
refactor the tools to use drag trackers directly in the following
commits.

The adapter will cause a dragging tool controller to render itself twice
until the controller is refactored to use a drag tracker directly.